broken out calculateUnsocialHours to its own function, and written simple test, using tests. Reformatted to make a notification folder, which will hold notif-specific tests, types, and helper functions.
