+++ src/app/constants.ts
@@ -1,1 +1,3 @@
 export const WEEKS_PER_YEAR = 52
+export const UNSOCIAL_HOUR_START = 21;
+export const UNSOCIAL_HOUR_END = 6;
