+++ src/app/notifications/helper-functions/geoHashHelpers.ts
@@ -1,5 +1,6 @@
+import { AugmentedParkingLocationData } from '@/parking-locations/types';
 import ngeohash from 'ngeohash';
-import { AugmentedParkingLocationData } from '../types';
+
 
 const geohashPrecision = 5; //each geohash represents an area approximately 4.9 km x 4.9 km with precision of 5.
